Split into db and storage
